layout/generic/test/test_bug385751.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=385751
-->
<head>
  <title>Test for Bug 385751</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <style>#content:first-letter {float: right; }
  </style>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=385751">Mozilla Bug 385751</a>
<div id="content"><span style="direction: rtl; unicode-bidi: embed;">*::first
m</span></div>
<pre id="test">
<script class="testbody" type="text/javascript">
 function do_Test()
{
/** Test for Bug 385751 **/
  ok(true, "Should not crash");
  SimpleTest.finish();
}
  SimpleTest.requestFlakyTimeout("untriaged");
  setTimeout(do_Test, 500);
  SimpleTest.waitForExplicitFinish();
</script>
</pre>
</body>
</html>
